+++ src/Infrastructure/Repository/HappeningParticipationRepository.php
@@ -9,6 +9,7 @@ use Proximum\Vimeet\Domain\Model\HappeningParticipation;
 use Proximum\Vimeet\Domain\Model\Participant;
 use Proximum\Vimeet\Domain\Model\Sheet;
 use Proximum\Vimeet\Domain\Model\User;
+use Proximum\Vimeet\Domain\Model\User\Event\Scan;
 use Proximum\Vimeet\Domain\Repository\HappeningParticipationRepositoryInterface;
 use Proximum\Vimeet\Domain\Scan\Type;
 
@@ -456,7 +457,7 @@ class HappeningParticipationRepository implements HappeningParticipationReposito
             ->select('participation')
             ->from(HappeningParticipation::class, 'participation')
             ->join('participation.happening', 'happening', 'WITH', 'happening.event = :event')
-            ->join(User\Event\Scan::class, 'scan', 'WITH', 'scan.type = :scan_type AND happening = scan.objectId AND scan.user = :user')
+            ->join(Scan::class, 'scan', 'WITH', 'scan.type = :scan_type AND happening = scan.objectId AND scan.user = :user')
             ->setParameter('event', $event)
             ->andWhere('participation.user = :user')
             ->setParameter('user', $user)
@@ -517,5 +518,24 @@ class HappeningParticipationRepository implements HappeningParticipationReposito
         }
 
         return $happeningEvaluationsAverage;
+    }
+
+    public function getScannedAt(HappeningParticipation $participation): ?\DateTimeInterface
+    {
+        $result = $this
+            ->entityManager
+            ->createQueryBuilder()
+            ->select('scan.scannedAt')
+            ->from(HappeningParticipation::class, 'participation')
+            ->join('participation.happening' , 'happening')
+            ->join(Scan::class, 'scan', 'WITH', 'scan.type = :scan_type AND happening.id = scan.objectId AND scan.user = participation.user')
+            ->andWhere('participation = :participation')
+            ->setParameter('participation', $participation)
+            ->setParameter('scan_type', Type::TYPE_HAPPENING_ENTRANCE)
+            ->setMaxResults(1)
+            ->getQuery()
+            ->getArrayResult();
+
+        return $result[0]['scannedAt'] ?? null;
     }
 }