+++ CHANGELOG.md
@@ -4,6 +4,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- MV-321 - webinar vote form not displayed
+
 ## [2.85.1] - 2021-03-30
 ### Fixed
 - MV-320  - check if possible slots are available