+++ CHANGELOG.md
@@ -4,10 +4,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.86.0] - 2021-03-31
 ### Added
 - VIMEET-2439 - Export conferences participants (phone / grade / connection date)
-
-### Added
 - VIMEET-2438 - Export conferences (Participant connected, Number of grades, Average grade)
 
 ## [2.85.1] - 2021-03-30